Add checks, add indexes, and fix response format.

+++ src/main/java/com/google/sps/servlets/BlobServlet.java
@@ -80,9 +80,9 @@ public class BlobServlet extends HttpServlet {
       }
     }
 
-    String uEmail = userService.getCurrentUser().getEmail();
+    String userEmail = userService.getCurrentUser().getEmail();
     String projId = request.getParameter("proj-id");
-    Entity projEntity = DataUtils.getProjectEntity(projId, uEmail, true, false);
+    Entity projEntity = DataUtils.getProjectEntity(projId, userEmail, true, false);
     Key projKey = projEntity.getKey();
     Key assetParentKey = projKey;
 
@@ -109,10 +109,11 @@ public class BlobServlet extends HttpServlet {
     // Owners have additional permissions
     ArrayList<String> owners =
         (ArrayList<String>)projEntity.getProperty("owners");
-    boolean isOwner = owners.contains(uEmail);
+    boolean isOwner = owners.contains(userEmail);
 
     if (isCreateMode && !isOwner && !isMask) {
-      throw new IOException("You do not have permission to to upload an image to this project.");
+      throw new IOException(
+          "You do not have permission to to upload an image to this project.");
     }
 
     boolean delete = Boolean.parseBoolean(request.getParameter("delete"));
@@ -130,30 +131,31 @@ public class BlobServlet extends HttpServlet {
         BlobstoreServiceFactory.getBlobstoreService();
     Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
     List<BlobKey> blobKeys = blobs.get("image");
-    BlobKey blobKey = blobKeys.get(0);
-    BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
-
-    // Check for blob size as well since no upload doesn't guarantee an empty
-    // BlobKey List
-    boolean hasNonEmptyImage =
-        !(blobKeys == null || blobKeys.isEmpty() || blobInfo.getSize() == 0);
 
     // User submitted form without selecting a file
-    if (!hasNonEmptyImage) {
+    if (blobKeys.isEmpty() || blobKeys == null) {
       if (isCreateMode) {
         throw new IOException("Form submitted without a file.");
       }
     } else {
+      BlobKey blobKey = blobKeys.get(0);
+      BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
+
+      // Check for blob size as well since (experimentally) no upload does not
+      // guarantee an empty BlobKey List
+      boolean hasNonEmptyImage = blobInfo.getSize() != 0;
+
+      // Image cannot be changed after first upload
       if (!isCreateMode && !isMask) {
         blobstoreService.delete(blobKey);
         hasNonEmptyImage = false;
       }
-    }
 
-    // Set blobkey property
-    if (isCreateMode || (!isCreateMode && hasNonEmptyImage)) {
-      checkFileValidity(blobKey, isMask);
-      imgEntity.setProperty("blobkey", blobKey.getKeyString());
+      // Set blobkey property
+      if (isCreateMode || (!isCreateMode && hasNonEmptyImage)) {
+        checkFileValidity(blobKey, isMask);
+        imgEntity.setProperty("blobkey", blobKey.getKeyString());
+      }
     }
 
     // Last-modified time
@@ -191,8 +193,8 @@ public class BlobServlet extends HttpServlet {
     // Send the image URL and name
     response.setContentType("application/json");
     String url =
-        "blob-host?blobkey=" + (String)imgEntity.getProperty("blobkey");
-    Gson gson = new GsonBuilder().setPrettyPrinting().create();
+        "/blob-host?blobkey=" + (String)imgEntity.getProperty("blobkey");
+    Gson gson = new GsonBuilder().setPrettyPrinting().disableHtmlEscaping().create();
     String jsonImgInfo = gson.toJson(new blobPostReturn(url, checkedName));
     response.getWriter().println(jsonImgInfo);
   }
@@ -212,7 +214,7 @@ public class BlobServlet extends HttpServlet {
     UserService userService = UserServiceFactory.getUserService();
     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
 
-    String uEmail = userService.getCurrentUser().getEmail();
+    String userEmail = userService.getCurrentUser().getEmail();
 
     // Must be logged in
     if (!userService.isUserLoggedIn()) {
@@ -221,7 +223,7 @@ public class BlobServlet extends HttpServlet {
     }
 
     String projId = request.getParameter("proj-id");
-    Entity projEntity = DataUtils.getProjectEntity(projId, uEmail, true, true);
+    Entity projEntity = DataUtils.getProjectEntity(projId, userEmail, true, true);
     Key projKey = projEntity.getKey();
 
     boolean withMasks =
@@ -269,7 +271,8 @@ public class BlobServlet extends HttpServlet {
                 .addSort("utc", sortMask.equals(DataUtils.ASCENDING_SORT)
                                     ? Query.SortDirection.ASCENDING
                                     : Query.SortDirection.DESCENDING)
-                .setFilter(combinedGetFilters(request, withMasks, DataUtils.MASK));
+                .setFilter(
+                    combinedGetFilters(request, withMasks, DataUtils.MASK));
 
         PreparedQuery storedMasks = datastore.prepare(maskQuery);
 
@@ -310,15 +313,25 @@ public class BlobServlet extends HttpServlet {
                  : Arrays.asList("png", "jpg", "jpeg", "jfif", "pjpeg", "pjp",
                                  "gif", "bmp", "ico", "cur", "svg", "webp"));
 
+    ArrayList<String> validMimeTypes = new ArrayList<String>(
+        (isMask) ? Arrays.asList("png")
+                 : Arrays.asList("image/png", "image/jpeg", "image/pjpeg",
+                                 "image/gif", "image/bmp", "image/x-icon",
+                                 "image/svg+xml", "image/webp"));
+
     BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
     String[] splitFilename = blobInfo.getFilename().split("\\.");
     String extension = splitFilename[splitFilename.length - 1].toLowerCase();
+    String mimeType = blobInfo.getContentType().toLowerCase();
+
     if (blobInfo.getSize() == 0) {
       blobstoreService.delete(blobKey);
       throw new IOException("Invalid Blobkey.");
-    } else if (!validExtensions.contains(extension)) {
+    } else if (!validExtensions.contains(extension) &&
+               !validMimeTypes.contains(mimeType)) {
       blobstoreService.delete(blobKey);
-      throw new IOException("File not supported.");
+      throw new IOException("File not supported; extension: " + extension +
+                            "; MIME type: " + mimeType);
     }
   }
 
@@ -353,14 +366,16 @@ public class BlobServlet extends HttpServlet {
    * @param     {String}                kind        either IMAGE or MASK
    * @return    {Filter}
    */
-  private Filter combinedGetFilters(HttpServletRequest request, boolean withMasks, String kind) {
+  private Filter combinedGetFilters(HttpServletRequest request,
+                                    boolean withMasks, String kind) {
     ArrayList<Filter> allImgFilters = new ArrayList<Filter>();
 
     String tag = request.getParameter("tag");
     boolean isImage = kind == DataUtils.IMAGE;
     boolean isTagApplicable = !(isImage && withMasks);
     if (!DataUtils.isEmptyParameter(tag) && isTagApplicable) {
-      Filter tagFilter = new FilterPredicate("tags", FilterOperator.EQUAL, tag.toLowerCase());
+      Filter tagFilter =
+          new FilterPredicate("tags", FilterOperator.EQUAL, tag.toLowerCase());
       allImgFilters.add(tagFilter);
     }
